refactor(runtime): re-extract runtime-types + trim builder/runtime-types JSDoc @examples
Two related cleanups closing the #76 follow-up:

1. **Re-extract MCPServerConfig / RuntimeOptions / LightRuntimeOptions** out
   of runtime.ts into runtime-types.ts (the W26-C extraction lost in the
   2026-05-25 merge resolution when the local full-file version had to win
   over the slim re-export form due to in-place additions). Re-exports
   preserved in runtime.ts for backward compatibility.

2. **Strip @example code blocks from builder.ts + runtime-types.ts JSDoc**.
   Doc-bulk reduction follow-up flagged in W26-B-2 retro. Inline one-line
   '@example .withFoo()' usage hints preserved; multi-line typescript code
   blocks removed. 60 + 13 = 73 @example blocks trimmed.

LOC results:
- runtime.ts: 2083 → 1258 (-825 / -40%, ≤1500 ✅)
- runtime-types.ts: new, 783 LOC
- builder.ts: 2345 → 1964 (-381 / -16%, still above 1500 but doc bulk
  remaining is narrative description, not extractable as type/logic)

Verification: workspace tests 5600/0/25-skip, build 38/38, runtime
package tests 831/0/1.
